Extract declaration lowering plan
